+++ README.md
@@ -1,1 +1,12 @@
 # Video_Streaming
+
+### Server
+
+- API calls will be made on `/api/{routes}`
+
+### Setting up
+
+1. git clone `https://github.com/Meet-Modi/Video_Streaming`
+2. `cd Video_Streaming`
+3. `npm install`
+4. `npm run start`
