+++ index.html
@@ -1,7 +1,7 @@
 <div class="row">
     <div class="col-md-3 col-md-offset-3">
         <video style="height: 400px; width:600px;" id="videoPlayer" controls muted="muted" autoplay>
-<source src="http://localhost:3000/video" type="video/mp4">
-</video>
+            <source src="http://localhost:3000/api/video" type="video/mp4">
+        </video>
     </div>
 </div>
