+++ routes/index.js
@@ -1,0 +1,10 @@
+const express = require('express')
+const path = require('path')
+
+const router = express.Router()
+
+router.get('/', function (req, res) {
+    res.sendFile(path.join(__dirname + '/..' + '/index.html'))
+})
+
+module.exports = router;
