+++ server.js
@@ -1,46 +1,25 @@
 const express = require('express')
-const fs = require('fs')
-const path = require('path')
+const bodyParser = require('body-parser')
+require('dotenv').config()
+
+// Router Files
+const videoRouter = require('./routes/video');
+const router = require('./routes');
+
+const env = process.env.NODE_ENV || 'DEVELOPMENT'
+
 const app = express()
+app.use(bodyParser.json())
 
-app.get('/', function(req, res) {
-    res.sendFile(path.join(__dirname + '/index.html'))
+app.get('/', (req, res) => {
+    res.send('Hello World!');
 })
 
-app.get('/video', function(req, res) {
-    const path = 'assets/sample.mp4'
-    const stat = fs.statSync(path)
-    const fileSize = stat.size
-    const range = req.headers.range
+app.use('/api', router);
+app.use('/api', videoRouter);
 
-    if (range) {
-        const parts = range.replace(/bytes=/, "").split("-")
-        const start = parseInt(parts[0], 10)
-        const end = parts[1] ?
-            parseInt(parts[1], 10) :
-            fileSize - 1
 
-        const chunksize = (end - start) + 1
-        const file = fs.createReadStream(path, { start, end })
-        const head = {
-            'Content-Range': `bytes ${start}-${end}/${fileSize}`,
-            'Accept-Ranges': 'bytes',
-            'Content-Length': chunksize,
-            'Content-Type': 'video/mp4',
-        }
-
-        res.writeHead(206, head)
-        file.pipe(res)
-    } else {
-        const head = {
-            'Content-Length': fileSize,
-            'Content-Type': 'video/mp4',
-        }
-        res.writeHead(200, head)
-        fs.createReadStream(path).pipe(res)
-    }
-})
-
-app.listen(3000, function() {
-    console.log('App is running on port 3000')
+const port = process.env.PORT || 3000
+app.listen(port, function () {
+    console.log('Server is running on port 3000')
 })
